Added gameobject interface, moved the horse handler
