+++ app/src/main/java/com/nhlstenden/appdev/features/casino/interfaces/GameObject.kt
@@ -1,0 +1,6 @@
+package com.nhlstenden.appdev.features.casino.interfaces
+
+interface GameObject {
+    fun start()
+    fun stop()
+}
