+++ app/src/main/java/com/nhlstenden/appdev/features/casino/models/GuineaHorseHandler.kt
@@ -1,8 +1,9 @@
-package com.nhlstenden.appdev.features.casino.fragments
+package com.nhlstenden.appdev.features.casino.models
 
 import android.os.Handler
 import android.os.Looper
 import android.widget.ImageView
+import com.nhlstenden.appdev.features.casino.interfaces.GameObject
 import com.nhlstenden.appdev.features.casino.interfaces.RaceManager
 import kotlin.random.Random
 
@@ -10,14 +11,14 @@ class GuineaHorseHandler(
     private val guineaHorse: ImageView,
     private val guineaHorseName: String,
     private val raceManager: RaceManager
-): Runnable {
+): Runnable, GameObject {
     val handler = Handler(Looper.getMainLooper())
 
-    fun start() {
+    override fun start() {
         handler.post(this)
     }
 
-    fun stop() {
+    override fun stop() {
         handler.removeCallbacksAndMessages(null)
     }
 
@@ -28,7 +29,7 @@ class GuineaHorseHandler(
 
             return
         }
-        val nextPosition = Random.nextInt(1, 15)
+        val nextPosition = Random.Default.nextInt(1, 15)
         guineaHorse.y += nextPosition.toFloat()
         handler.postDelayed(this, nextPosition.toLong())
     }
